ref: Move SQL instrumentation out of Django
